Cleanup of useless files
