+++ back/src/Model/Websocket/CharacterLayer.ts
@@ -1,0 +1,4 @@
+export interface CharacterLayer {
+    name: string,
+    url: string|undefined
+}
