+++ back/src/Model/Websocket/ProtobufUtils.ts
@@ -5,7 +5,7 @@ import {
     PointMessage,
     PositionMessage
 } from "../../Messages/generated/messages_pb";
-import {CharacterLayer, ExSocketInterface} from "_Model/Websocket/ExSocketInterface";
+import {CharacterLayer} from "_Model/Websocket/CharacterLayer";
 import Direction = PositionMessage.Direction;
 import {ItemEventMessageInterface} from "_Model/Websocket/ItemEventMessage";
 import {PositionInterface} from "_Model/PositionInterface";
