--- pusher/src/Model/Distance.ts
@@ -1,7 +1,0 @@
-import {MessageUserPosition} from "../Model/Websocket/MessageUserPosition";
-
-export interface Distance {
-    distance: number,
-    first: MessageUserPosition,
-    second: MessageUserPosition,
-}
