--- pusher/src/Model/Websocket/GroupUpdateInterface.ts
@@ -1,6 +1,0 @@
-import {PositionInterface} from "_Model/PositionInterface";
-
-export interface GroupUpdateInterface {
-    position: PositionInterface,
-    groupId: number,
-}
