+++ pusher/src/Model/Websocket/Point.ts
@@ -4,8 +4,3 @@ export class Point implements PointInterface{
     constructor(public x : number, public y : number, public direction : string = "none", public moving : boolean = false) {
     }
 }
-
-export class MessageUserPosition {
-    constructor(public userId: number, public name: string, public characterLayers: string[], public position: PointInterface) {
-    }
-}
