--- pusher/src/Model/Websocket/SetPlayerDetailsMessage.ts
@@ -1,8 +1,0 @@
-import * as tg from "generic-type-guard";
-
-export const isSetPlayerDetailsMessage =
-    new tg.IsInterface().withProperties({
-        name: tg.isString,
-        characterLayers: tg.isArray(tg.isString)
-    }).get();
-export type SetPlayerDetailsMessage = tg.GuardedType<typeof isSetPlayerDetailsMessage>;
