--- pusher/src/Model/Websocket/UserInGroupInterface.ts
@@ -1,5 +1,0 @@
-export interface UserInGroupInterface {
-    userId: number,
-    name: string,
-    initiator: boolean
-}
